MDRange: Added perf test to Cuda, minor fixes

+++ core/src/Cuda/KokkosExp_Cuda_IterateTile.hpp
@@ -1144,8 +1144,8 @@ public:
       // TODO: Check product of tiles fits within max size < 1024
       const dim3 block( m_rp.m_tile[0]*m_rp.m_tile[1] , m_rp.m_tile[2] , m_rp.m_tile[3] ); //pad for mult of 16? check within max num threads bounds? 
       const dim3 grid( 
-          std::min( ( ( m_rp.m_upper[0] - m_rp.m_lower[0] + m_rp.m_tile[0] - 1 ) / m_rp.m_tile[0] 
-                    *  ( m_rp.m_upper[1] - m_rp.m_lower[1] + m_rp.m_tile[1] - 1 ) / m_rp.m_tile[1] )
+          std::min( ( static_cast<int>(( m_rp.m_upper[0] - m_rp.m_lower[0] + m_rp.m_tile[0] - 1 ) / m_rp.m_tile[0] 
+                    *  ( m_rp.m_upper[1] - m_rp.m_lower[1] + m_rp.m_tile[1] - 1 ) / m_rp.m_tile[1] ))
                   , static_cast<int>(maxblocks) ) 
           //std::min( ( m_rp.m_upper[0] - m_rp.m_lower[0] + block.x - 1 ) / block.x , maxblocks ) 
         , std::min( ( m_rp.m_upper[2] - m_rp.m_lower[2] + block.y - 1 ) / block.y , maxblocks ) 
@@ -1159,11 +1159,11 @@ public:
       // TODO: Check product of tiles fits within max size < 1024
       const dim3 block( m_rp.m_tile[0]*m_rp.m_tile[1] , m_rp.m_tile[2]*m_rp.m_tile[3] , m_rp.m_tile[4] ); //pad for mult of 16? check within max num threads bounds? 
       const dim3 grid( 
-          std::min( ( ( m_rp.m_upper[0] - m_rp.m_lower[0] + m_rp.m_tile[0] - 1 ) / m_rp.m_tile[0] 
-                   *  ( m_rp.m_upper[1] - m_rp.m_lower[1] + m_rp.m_tile[1] - 1 ) / m_rp.m_tile[1] )
+          std::min( static_cast<int>(( ( m_rp.m_upper[0] - m_rp.m_lower[0] + m_rp.m_tile[0] - 1 ) / m_rp.m_tile[0] 
+                   *  ( m_rp.m_upper[1] - m_rp.m_lower[1] + m_rp.m_tile[1] - 1 ) / m_rp.m_tile[1] ))
                   , static_cast<int>(maxblocks) ) 
-        , std::min( ( ( m_rp.m_upper[2] - m_rp.m_lower[2] + m_rp.m_tile[2] - 1 ) / m_rp.m_tile[2] 
-                   *  ( m_rp.m_upper[3] - m_rp.m_lower[3] + m_rp.m_tile[3] - 1 ) / m_rp.m_tile[3] )
+        , std::min( static_cast<int>(( ( m_rp.m_upper[2] - m_rp.m_lower[2] + m_rp.m_tile[2] - 1 ) / m_rp.m_tile[2] 
+                   *  ( m_rp.m_upper[3] - m_rp.m_lower[3] + m_rp.m_tile[3] - 1 ) / m_rp.m_tile[3] ))
                   , static_cast<int>(maxblocks) ) 
         //, std::min( ( m_rp.m_upper[2] - m_rp.m_lower[2] + block.y - 1 ) / block.y , maxblocks ) 
         , std::min( ( m_rp.m_upper[4] - m_rp.m_lower[4] + block.z - 1 ) / block.z , maxblocks ) 
@@ -1176,14 +1176,14 @@ public:
       // TODO: Check product of tiles fits within max size < 1024
       const dim3 block( m_rp.m_tile[0]*m_rp.m_tile[1] , m_rp.m_tile[2]*m_rp.m_tile[3] , m_rp.m_tile[4]*m_rp.m_tile[5] ); //pad for mult of 16? check within max num threads bounds? 
       const dim3 grid( 
-          std::min( ( ( m_rp.m_upper[0] - m_rp.m_lower[0] + m_rp.m_tile[0] - 1 ) / m_rp.m_tile[0] 
-                   *  ( m_rp.m_upper[1] - m_rp.m_lower[1] + m_rp.m_tile[1] - 1 ) / m_rp.m_tile[1] )
+          std::min( static_cast<int>(( ( m_rp.m_upper[0] - m_rp.m_lower[0] + m_rp.m_tile[0] - 1 ) / m_rp.m_tile[0] 
+                   *  ( m_rp.m_upper[1] - m_rp.m_lower[1] + m_rp.m_tile[1] - 1 ) / m_rp.m_tile[1] ))
                   , static_cast<int>(maxblocks) ) 
-        , std::min( ( ( m_rp.m_upper[2] - m_rp.m_lower[2] + m_rp.m_tile[2] - 1 ) / m_rp.m_tile[2] 
-                   *  ( m_rp.m_upper[3] - m_rp.m_lower[3] + m_rp.m_tile[3] - 1 ) / m_rp.m_tile[3] )
+        , std::min( static_cast<int>(( ( m_rp.m_upper[2] - m_rp.m_lower[2] + m_rp.m_tile[2] - 1 ) / m_rp.m_tile[2] 
+                   *  ( m_rp.m_upper[3] - m_rp.m_lower[3] + m_rp.m_tile[3] - 1 ) / m_rp.m_tile[3] ))
                   , static_cast<int>(maxblocks) ) 
-        , std::min( ( ( m_rp.m_upper[4] - m_rp.m_lower[4] + m_rp.m_tile[4] - 1 ) / m_rp.m_tile[4] 
-                   *  ( m_rp.m_upper[5] - m_rp.m_lower[5] + m_rp.m_tile[5] - 1 ) / m_rp.m_tile[5] )
+        , std::min( static_cast<int>(( ( m_rp.m_upper[4] - m_rp.m_lower[4] + m_rp.m_tile[4] - 1 ) / m_rp.m_tile[4] 
+                   *  ( m_rp.m_upper[5] - m_rp.m_lower[5] + m_rp.m_tile[5] - 1 ) / m_rp.m_tile[5] ))
                   , static_cast<int>(maxblocks) ) 
         );
       CudaLaunch< DeviceIterateTile >( *this , grid , block );
